+++ cats/signals/types.py
@@ -52,4 +52,5 @@ class DomainProvenanceResult(SignalResult):
     free_host: bool = False
     typosquat: bool = False
     brand_on_bad_tld: bool = False
+    low_popularity_corroboration: bool = False
     host: str = ""